+++ src/squaremap_combine/combine_cli.py
@@ -17,10 +17,10 @@ from squaremap_combine.combine_core import (
     CombinerStyle,
     logger,
 )
-from squaremap_combine.helper import confirm_yn, filled_tuple
 from squaremap_combine.logging import enable_logging
 from squaremap_combine.project import LOGS_DIR, PROJECT_VERSION
 from squaremap_combine.type_alias import Rectangle
+from squaremap_combine.util import confirm_yn, filled_tuple
 
 
 def opt(*names: str) -> list[str]: