+++ src/squaremap_combine/const.py
@@ -1,0 +1,8 @@
+from collections import OrderedDict
+
+SQMAP_DETAIL_LEVELS: OrderedDict[int, float] = OrderedDict({
+    0: 1.0,
+    1: 1 / 2,
+    2: 1 / 4,
+    3: 1 / 8,
+})