+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added module `const`
+
 ### Changed
 
 - In `combine_core`:
@@ -17,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - All options now either only use a single character for their short name, or have no short name at all
   - `--output-dir` renamed to `--out`, now takes a file path instead of a directory path
   - `-a/--area` renamed to `-r/--rect`
+- Renamed module `helper` to `util`
 
 ### Deprecated